round last_reported to nearest hour, skip non-reporting stations
Stations rarely report once per hour so round to nearst hour.  This helps reporting so that we can get system-wide hourly numbers.
Stations with last_reported=0 seem to not be reporting.  Skip those.  Might be good to set last_reported=now() and report zero stats so that we can track non-reporting stations.